Got rid of extra file, renamed repoDatabaseNav to repoLeftNav
